requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
openai
redis
aiohttp
httpx
numpy
